# openapi.yaml
openapi: 3.0.3
info:
  title: Lingora API
  version: 0.1.0
  description: |
    API nội dung học tiếng Anh của Lingora.

    Route đọc là công khai. Route ghi cần access token JWT HS256 có claim
    `role: admin`.
servers:
  - url: http://localhost:8080/v1
    description: Máy dev
  - url: https://api.lingora.vn/v1
    description: Production
...
paths:
...
  /courses:
    get:
      tags: [courses]
      summary: Liệt kê khoá học
      description: |
        Phân trang keyset. Trang đầu không truyền `cursor`; các trang sau truyền
        lại nguyên văn `next_cursor` của trang trước. `next_cursor` bằng `null`
        nghĩa là đã hết dữ liệu.
      operationId: listCourses
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/LevelFilter'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Một trang khoá học
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [courses]
      summary: Tạo khoá học
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Khoá học đã được tạo
          headers:
            Location:
              description: Đường dẫn tới khoá học vừa tạo
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /courses/order:
    put:
      tags: [courses]
      summary: Sắp xếp khoá học trong một bậc
      description: |
        `course_ids` phải là đúng tập khoá hiện có của bậc đó. Thiếu, thừa hay
        trùng đều bị từ chối, vì một lần sắp xếp thiếu khoá sẽ để lại những khoá
        mang thứ tự cũ lẫn vào giữa.
      operationId: reorderCourses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseOrder'
      responses:
        '204':
          description: Đã sắp xếp
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{courseId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'

    get:
      tags: [courses]
      summary: Lấy khoá học theo id
      description: |
        Khoá `draft` chỉ admin đọc được; người khác nhận 404 chứ không phải 403,
        vì 403 đã xác nhận bản ghi đó tồn tại.
      operationId: getCourse
      responses:
        '200':
          description: Khoá học
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [courses]
      summary: Cập nhật một phần khoá học
      description: |
        Field vắng mặt thì giữ nguyên. Riêng `cover_image_url` gửi `null` nghĩa
        là xoá ảnh bìa. Body rỗng bị từ chối.
      operationId: updateCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Khoá học sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [courses]
      summary: Xoá mềm khoá học
      operationId: deleteCourse
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Đã xoá
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/by-slug/{slug}:
    get:
      tags: [courses]
      summary: Lấy khoá học theo slug
      description: |
        Khoá `draft` chỉ admin đọc được; người khác nhận 404 chứ không phải 403,
        vì 403 đã xác nhận bản ghi đó tồn tại.
      operationId: getCourseBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: ngu-phap-co-ban
      responses:
        '200':
          description: Khoá học
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/lessons:
    parameters:
      - $ref: '#/components/parameters/CourseId'

    get:
      tags: [courses]
      summary: Liệt kê bài học của một khoá
      description: |
        Sắp xếp theo `position` tăng dần. Khoá không tồn tại trả về 404; khoá có
        thật nhưng chưa có bài trả về mảng rỗng.

        Bài không có trạng thái riêng, nó thừa hưởng của khoá: bài thuộc khoá
        `draft` chỉ admin đọc được, người khác cũng nhận 404.
      operationId: listCourseLessons
      responses:
        '200':
          description: Danh sách bài học
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [courses]
      summary: Thêm bài học vào khoá
      description: Bài mới luôn đứng cuối; đổi thứ tự bằng endpoint order.
      operationId: createLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreate'
      responses:
        '201':
          description: Bài học đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /courses/{courseId}/lessons/order:
    parameters:
      - $ref: '#/components/parameters/CourseId'

    put:
      tags: [courses]
      summary: Sắp xếp lại bài học trong khoá
      description: |
        Danh sách phải liệt kê đúng tập bài hiện có của khoá — thiếu, thừa hay
        trùng đều bị từ chối. Vị trí mới lấy từ chính thứ tự của mảng.
      operationId: reorderLessons
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonOrder'
      responses:
        '204':
          description: Đã sắp xếp lại
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lessons/{lessonId}:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          x-go-type: string

    get:
      tags: [courses]
      summary: Lấy bài học theo id
      description: |
        Bài không có trạng thái riêng, nó thừa hưởng của khoá: bài thuộc khoá
        `draft` chỉ admin đọc được, người khác nhận 404.
      operationId: getLesson
      responses:
        '200':
          description: Bài học kèm nội dung
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [courses]
      summary: Cập nhật một phần bài học
      description: |
        position không sửa được ở đây; dùng endpoint order của khoá.
      operationId: updateLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdate'
      responses:
        '200':
          description: Bài học sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [courses]
      summary: Xoá mềm bài học
      operationId: deleteLesson
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Đã xoá
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lessons/{lessonId}/blocks:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          x-go-type: string

    put:
      tags: [courses]
      summary: Thay toàn bộ nội dung của một bài
      description: |
        Nhận cả danh sách chứ không sửa từng khối: soạn bài là việc kéo thả và
        viết lại, nên gửi trọn trạng thái mong muốn đơn giản hơn nhiều so với
        một chuỗi thao tác thêm/sửa/xoá/đổi chỗ.

        `position` không có trong body — thứ tự chính là thứ tự của mảng.
      operationId: replaceLessonBlocks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonBlockList'
      responses:
        '204':
          description: Đã thay nội dung
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token HS256 với claim `role: admin`."
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        x-go-type: string
    StatusFilter:
      name: status
      in: query
      required: false
      description: |
        Chỉ có tác dụng với token admin. Người gọi khác luôn nhận `published`,
        và nếu xin `draft` thì nhận một trang rỗng.
      schema:
        $ref: '#/components/schemas/CourseStatus'
    LevelFilter:
      name: level
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/CourseLevel'
    PageSize:
      name: page_size
      in: query
      required: false
      description: Mặc định 20, trần 100.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
    Cursor:
      name: cursor
      in: query
      required: false
      description: Chuỗi mờ lấy từ `next_cursor` của trang trước.
      schema:
        type: string
  schemas:
...
    VocabularyEntry:
      type: object
      required: [id, lesson_id, word, ipa, meaning, example, example_vi, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        lesson_id:
          type: string
          format: uuid
          x-go-type: string
        word:
          type: string
          maxLength: 100
        ipa:
          type: string
          maxLength: 100
        meaning:
          type: string
          maxLength: 300
        example:
          type: string
          maxLength: 500
        example_vi:
          type: string
          maxLength: 500
        position:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CourseLevel:
      type: string
      description: Trình độ theo khung CEFR.
      enum: [A1, A2, B1, B2, C1, C2]
    CourseStatus:
      type: string
      description: Trạng thái xuất bản.
      enum: [draft, published]
    Course:
      type: object
      required: [id, slug, title, description, level, status, cover_image_url, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        slug:
          type: string
          example: ngu-phap-co-ban
        title:
          type: string
          example: Ngữ pháp cơ bản
        description:
          type: string
        level:
          $ref: '#/components/schemas/CourseLevel'
        status:
          $ref: '#/components/schemas/CourseStatus'
        cover_image_url:
          type: string
          format: uri
          nullable: true
        position:
          type: integer
          format: int32
          description: Thứ tự trong bậc, do người soạn đặt.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CourseList:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        next_cursor:
          type: string
          nullable: true
          description: Con trỏ cho trang kế tiếp; `null` nghĩa là đã hết.
    CourseCreate:
      type: object
      required: [slug, title, level]
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        level:
          $ref: '#/components/schemas/CourseLevel'
        status:
          $ref: '#/components/schemas/CourseStatus'
        cover_image_url:
          type: string
          format: uri
          maxLength: 2048
          nullable: true
    CourseOrder:
      type: object
      required: [level, course_ids]
      properties:
        level:
          $ref: '#/components/schemas/CourseLevel'
        course_ids:
          type: array
          description: Toàn bộ id khoá của bậc đó, theo thứ tự mong muốn.
          items:
            type: string
            format: uuid
            x-go-type: string
    CourseUpdate:
      type: object
      minProperties: 1
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        level:
          $ref: '#/components/schemas/CourseLevel'
        status:
          $ref: '#/components/schemas/CourseStatus'
        cover_image_url:
          type: string
          format: uri
          maxLength: 2048
          nullable: true
    Lesson:
      type: object
      required: [id, course_id, slug, title, summary, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        course_id:
          type: string
          format: uuid
          x-go-type: string
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
          description: Đoạn dẫn ngắn hiện ngay dưới tiêu đề bài; có thể rỗng.
        position:
          type: integer
          format: int32
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LessonBlockKind:
      type: string
      enum: [note, example, dialogue]
      description: |
        `note` là đoạn giải thích tiếng Việt, `example` là câu mẫu Anh–Việt,
        `dialogue` giống `example` nhưng có tên người nói.
    LessonBlock:
      type: object
      required: [kind, body, text_en, text_vi, speaker]
      description: |
        Mỗi dạng chỉ dùng một phần các trường, phần còn lại phải là chuỗi rỗng:

        | kind | trường bắt buộc | phải rỗng |
        | --- | --- | --- |
        | `note` | `body` | `text_en`, `text_vi`, `speaker` |
        | `example` | `text_en` | `body`, `speaker` |
        | `dialogue` | `text_en`, `speaker` | `body` |

        Database có ràng buộc bắt đúng hình dạng này, nên gửi sai là 400 chứ
        không phải một khối trống giữa bài.
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
          description: Chỉ có khi đọc; lúc ghi thì bỏ qua vì cả danh sách được thay mới.
        kind:
          $ref: '#/components/schemas/LessonBlockKind'
        body:
          type: string
        text_en:
          type: string
        text_vi:
          type: string
        speaker:
          type: string
    LessonBlockList:
      type: object
      required: [blocks]
      description: |
        Thay TOÀN BỘ nội dung của bài. Gửi mảng rỗng là xoá sạch nội dung — đó
        là cách duy nhất để làm việc đó, và nó có chủ đích.
      properties:
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/LessonBlock'
    LessonDetail:
      allOf:
        - $ref: '#/components/schemas/Lesson'
        - type: object
          required: [blocks, vocabulary]
          properties:
            blocks:
              type: array
              description: Nội dung bài theo đúng thứ tự hiển thị.
              items:
                $ref: '#/components/schemas/LessonBlock'
            vocabulary:
              type: array
              description: |
                Từ vựng bài này dạy, theo thứ tự của người soạn. Trả cho mọi
                người đọc được bài — khác với /me/vocabulary, vốn chỉ gồm từ của
                những bài đã học xong.
              items:
                $ref: '#/components/schemas/VocabularyEntry'
    LessonCreate:
      type: object
      required: [slug, title]
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
    LessonUpdate:
      type: object
      minProperties: 1
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 1000
    LessonOrder:
      type: object
      required: [lesson_ids]
      properties:
        lesson_ids:
          type: array
          minItems: 1
          description: Toàn bộ id bài của khoá, theo thứ tự mong muốn.
          items:
            type: string
            format: uuid
            x-go-type: string
    LessonList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - validation_error
            - malformed_body
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - conflict
            - rate_limited
            - internal_error
        message:
          type: string
          description: Câu thông báo tiếng Việt, hiển thị được cho người dùng.
        details:
          type: object
          additionalProperties:
            type: string
          description: Lỗi theo từng field, chỉ có với validation_error.
...
  responses:
...
    BadRequest:
      description: Dữ liệu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Thiếu token hoặc token không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Token hợp lệ nhưng không đủ quyền
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Không tìm thấy
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Xung đột với dữ liệu đang có
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'